requirements.txt
chromadb==0.4.0
uvicorn[standard]==0.23.2
fastapi==0.99.1
pydantic==1.10.9
sentence-transformers==2.2.2
cohere==4.9.0
numpy==1.24.4
scipy==1.10.1
sqlalchemy==2.0.21
hnswlib==0.7.0
click==8.1.7
duckdb==0.8.1
python-dotenv==1.0.0
torch==2.0.1  # Necesario para sentence-transformers
transformers==4.34.0  # Requerido por sentence-transformers
python-docx==0.8.11
